Bug 571334: [Diagram][CleanUp] Clean code for UML Diagrams   - update classpath (just to be sure)   - reapply papyrus code template, formatter and clean up   - format code   - organize import   - clean up

+++ plugins/uml/diagram/org.eclipse.papyrus.uml.diagram.common/src/org/eclipse/papyrus/uml/diagram/common/commands/PreserveAnchorsPositionCommand.java
@@ -261,20 +261,20 @@ public class PreserveAnchorsPositionCommand extends AbstractTransactionalCommand
 		Rectangle figureBounds = getFigureBounds();
 		PrecisionPoint pp = BaseSlidableAnchor.parseTerminalString(anchor.getId());
 		int figureHeight = figureBounds.height;
-		if (0!=figureHeight) {
+		if (0 != figureHeight) {
 			if (null == pp) {
 				pp = new PrecisionPoint();
 			}
 
 			if (PRESERVE_Y == getPreserveAxis() || PRESERVE_XY == getPreserveAxis()) {
-				//int anchorYPos = (int) Math.round(figureBounds.height * pp.preciseY());
+				// int anchorYPos = (int) Math.round(figureBounds.height * pp.preciseY());
 				double anchorYPos = figureHeight * pp.preciseY();
-				pp.setPreciseY( anchorYPos / (double)(figureHeight + sizeDelta.height));
+				pp.setPreciseY(anchorYPos / (figureHeight + sizeDelta.height));
 
 				// If the resize direction is NORTH, the location of the figure
 				// move, but the anchor stay visually at the same location
 				if (PositionConstants.NORTH == resizeDirection || PositionConstants.NORTH_EAST == resizeDirection || PositionConstants.NORTH_WEST == resizeDirection) {
-					pp.setPreciseY(pp.preciseY() + ((double) sizeDelta.height / (double)(figureHeight + sizeDelta.height)));
+					pp.setPreciseY(pp.preciseY() + ((double) sizeDelta.height / (double) (figureHeight + sizeDelta.height)));
 				}
 
 				if (pp.preciseY() > 1.0) {
@@ -285,14 +285,14 @@ public class PreserveAnchorsPositionCommand extends AbstractTransactionalCommand
 			}
 
 			if (PRESERVE_X == getPreserveAxis() || PRESERVE_XY == getPreserveAxis()) {
-				//int anchorXPos = (int) Math.round(figureBounds.width * pp.preciseX());
+				// int anchorXPos = (int) Math.round(figureBounds.width * pp.preciseX());
 				double anchorXPos = figureBounds.width * pp.preciseX();
-				pp.setPreciseX( anchorXPos / (double)(figureBounds.width + sizeDelta.width));
+				pp.setPreciseX(anchorXPos / (figureBounds.width + sizeDelta.width));
 
 				// If the resize direction is WEST, the location of the figure move,
 				// but the anchor stay visually at the same location
 				if (PositionConstants.WEST == resizeDirection || PositionConstants.NORTH_WEST == resizeDirection || PositionConstants.SOUTH_WEST == resizeDirection) {
-					pp.setPreciseX(pp.preciseX() + ((double) sizeDelta.width / (double)(figureBounds.width + sizeDelta.width)));
+					pp.setPreciseX(pp.preciseX() + ((double) sizeDelta.width / (double) (figureBounds.width + sizeDelta.width)));
 				}
 
 				if (pp.preciseX() > 1.0) {
